chore: Update requirements.txt and server.py
This commit updates the requirements.txt file to include the dependencies "duckduckgo-search" and "langchainhub". It also adds the "vocode" dependency to the requirements.txt file. Additionally, the server.py file is modified to import the "call_phone_number" function from the "outbound_call" module and the "get_all_contacts" function from the "contacts" module.

Note: The commit message has been modified to adhere to the established commit message conventions in the repository.