chore: Update contacts.py to include email field in CONTACTS
This commit updates the contacts.py file to include an "email" field in the CONTACTS list. The email address is set to "divesh.chowdary@gmail.com" for the "Test Venue" contact. This change allows for storing and accessing email information for contacts in the application.

Note: The commit message has been modified to adhere to the established commit message conventions in the repository.